CLN: Deleted deprecated plotting code.
